15. Now we update the value of the multiSelect, so we can select values and are updated in the commaSeparated and multiline.
We need the selectedOptions property selectedOptions. Ath teh begining
like we are not updating the state of the multiselect, the empty array
is re-render and thats why we are not seeing the selections. Because the
re-render is triggered every time the state is updated and is assigned
to a element, so this element is updated only (is more complicated than
that, for for now we say this).